+++ scripts/tweak
