+++ src/plugins/mcode/Magnetic_Bodge.m
