+++ src/plugins/mcode/Theory_Visual.m
