+++ src/plugins/mcode/__init__.py
@@ -1,0 +1,1 @@
+# backend package loader file - placeholder
