+++ tex/doc/doc.tex
@@ -47,7 +47,7 @@ sorting=nty
 
 \renewcommand{\abstractname}{Summary}
 
-\title{solensim v2.0.0-alpha project documentation - WIP}
+\title{solensim v0.2.1 project documentation - WIP}
 
 
 \author{Anton Douginets (anton.douginets@physik.hu-berlin.de)
